refactor(labelstudio): move the LS export builders into the library (T051b)
Follows the architecture review's T051b: the Label Studio export was
library-grade code sitting in the CLI, while the uncertainty-track builders it
sits beside already lived in labelstudio.py. Ten functions (~289 lines) move, so
the whole LS surface is now in one module. Script 1730 → 1491 lines; only the two
public entry points (build_labelstudio_task / build_labelstudio_config) are
imported back.

**Found a second divergent duplicate — this one had a real consequence.**
`_asr_has_timestamps` existed in *both* the script and harvesters.py, with
different semantics:

- harvesters (strict): needs a non-null `start` on a chunk or the line.
- script (loose): `start is not None or len(chunks) > 0`.

They disagree on a chunked-but-untimed transcript: the loose version calls it
timestamped. `stage_alignment` used the loose one to decide whether to skip
alignment — so exactly the input alignment exists to fix would be skipped and
never get timestamps. Meanwhile `resolve_asr_result`, two functions away in
harvesters.py, has always used the strict version, so the codebase disagreed with
itself about the same transcript.

Consolidated on the strict semantics, with the reasoning in the docstring and a
test pinning the disagreement case. Behavior change: a chunked-but-untimed
transcript is now aligned instead of silently skipped. That's the intended
direction, but it is a change.

`seg_attr` was also duplicated — byte-identical this time, so dropping the copy
is a pure cleanup.

Two script tests that reached for the moved privates were relocated with the code
rather than kept alive by re-export shims.

Verified: 403 passed / 15 skipped; mypy clean (286 files); ruff + format clean.
End-to-end re-run against the pre-move baseline: labelstudio_tasks.json and
labelstudio_config.xml byte-identical, uncertainty parquets byte-identical (3/3).
