+++ scripts/analyze_audio.py
@@ -147,6 +147,10 @@ from senselab.audio.workflows.audio_analysis.harvesters import (
 from senselab.audio.workflows.audio_analysis.harvesters import (
     classification_windows as _classification_windows,
 )
+from senselab.audio.workflows.audio_analysis.labelstudio import (
+    build_labelstudio_config,
+    build_labelstudio_task,
+)
 from senselab.audio.workflows.audio_analysis.stage_context import STAGE_VERSIONS, PassPlan, StageContext
 from senselab.audio.workflows.audio_analysis.stages import (
     _asr_has_timestamps,
@@ -803,315 +807,6 @@ def prepare_audio(path: Path) -> Audio:
 # (see ``cache_key``) and stamped into parquet provenance via
 # ``_CACHE_SCHEMA_VERSION`` references — never hardcode the literal anywhere
 # else, otherwise the constant and the stamped value will drift.
-
-
-def _new_region_id(prefix: str, idx: int) -> str:
-    """Stable per-region ID for Label Studio result entries."""
-    return f"{prefix}_{idx:04d}"
-
-
-def _ls_label_region(
-    *,
-    region_id: str,
-    from_name: str,
-    start: float,
-    end: float,
-    label: str,
-    score: float | None = None,
-) -> dict[str, Any]:
-    """Build one Label Studio ``labels`` result entry on the audio timeline."""
-    value: dict[str, Any] = {"start": float(start), "end": float(end), "labels": [label]}
-    entry: dict[str, Any] = {
-        "id": region_id,
-        "from_name": from_name,
-        "to_name": "audio",
-        "type": "labels",
-        "value": value,
-    }
-    if score is not None:
-        entry["score"] = float(score)
-    return entry
-
-
-def _ls_textarea_region(
-    *,
-    region_id: str,
-    from_name: str,
-    start: float,
-    end: float,
-    text: str,
-) -> dict[str, Any]:
-    """Build one Label Studio ``textarea`` per-region transcription entry."""
-    return {
-        "id": region_id,
-        "from_name": from_name,
-        "to_name": "audio",
-        "type": "textarea",
-        "value": {"start": float(start), "end": float(end), "text": [text]},
-    }
-
-
-def _seg_attr(seg: Any, name: str) -> Any:  # noqa: ANN401
-    """Return ``seg.name`` whether ``seg`` is a Pydantic model or a JSON dict.
-
-    Cache reads deserialize ScriptLine into plain dicts; in-memory results are
-    Pydantic objects. Both shapes flow through the LS helpers.
-    """
-    if isinstance(seg, dict):
-        return seg.get(name)
-    return getattr(seg, name, None)
-
-
-def _diarization_to_ls(result: Any, prefix: str) -> list[dict[str, Any]]:  # noqa: ANN401
-    """Convert diarize_audios output (List[List[ScriptLine]]) into LS regions."""
-    out: list[dict[str, Any]] = []
-    if not result:
-        return out
-    segments = result[0] if isinstance(result, list) and result else []
-    for i, seg in enumerate(segments):
-        start = _seg_attr(seg, "start")
-        end = _seg_attr(seg, "end")
-        speaker = _seg_attr(seg, "speaker") or "SPEAKER_UNKNOWN"
-        if start is None or end is None:
-            continue
-        out.append(
-            _ls_label_region(
-                region_id=_new_region_id(f"{prefix}_dia", i),
-                from_name=prefix,
-                start=start,
-                end=end,
-                label=str(speaker),
-            )
-        )
-    return out
-
-
-def _classification_to_ls(
-    result: Any,  # noqa: ANN401
-    prefix: str,
-    win_length: float,
-    hop_length: float,
-) -> list[dict[str, Any]]:
-    """Convert classify_audios windowed output into LS regions (top-1 per window).
-
-    Window centers advance by ``hop_length`` and span ``win_length`` seconds, so
-    the LS regions reflect each model's own native frame stride.
-    """
-    out: list[dict[str, Any]] = []
-    windows = _classification_windows(result)
-    for i, window in enumerate(windows):
-        label, score, _entropy = _classification_window_top1(window)
-        if label is None:
-            continue
-        # Prefer per-window start/end when the canonical shape carries them.
-        if isinstance(window, dict) and window.get("start") is not None and window.get("end") is not None:
-            start = float(window["start"])
-            end = float(window["end"])
-        else:
-            start = i * hop_length
-            end = start + win_length
-        out.append(
-            _ls_label_region(
-                region_id=_new_region_id(f"{prefix}_cls", i),
-                from_name=prefix,
-                start=start,
-                end=end,
-                label=label,
-                score=score if score is not None else 0.0,
-            )
-        )
-    return out
-
-
-def _asr_to_ls(result: Any, prefix: str, full_duration: float) -> list[dict[str, Any]]:  # noqa: ANN401
-    """Convert transcribe_audios output into LS textarea regions, one per ScriptLine.
-
-    Whisper sometimes returns one ScriptLine without timing for a short clip; in
-    that case we pin the textarea to the full audio span.
-    """
-    out: list[dict[str, Any]] = []
-    if not result:
-        return out
-    lines = result if isinstance(result, list) else [result]
-    for i, line in enumerate(lines):
-        text = _seg_attr(line, "text") or ""
-        start = _seg_attr(line, "start")
-        end = _seg_attr(line, "end")
-        if start is None or end is None:
-            start, end = 0.0, full_duration
-        if not text:
-            continue
-        out.append(
-            _ls_textarea_region(
-                region_id=_new_region_id(f"{prefix}_asr", i),
-                from_name=prefix,
-                start=start,
-                end=end,
-                text=text,
-            )
-        )
-    return out
-
-
-def build_labelstudio_task(
-    audio_uri: str,
-    pass_label: str,
-    duration_s: float,
-    pass_summary: dict[str, Any],
-    ast_win_length: float,
-    ast_hop_length: float,
-    yamnet_win_length: float,
-    yamnet_hop_length: float,
-) -> dict[str, Any]:
-    """Build one Label Studio task with predictions for all analyzers in this pass.
-
-    Each analyzer (diarization, ast, yamnet, asr) becomes its own
-    ``from_name`` track on the audio timeline, so the importer sees parallel
-    annotation rows. When a senselab task was run with multiple models, every
-    model's output is exported as its own track (e.g., ``asr_whisper_turbo``,
-    ``asr_whisper_small``) so they can be visually compared.
-    """
-    regions: list[dict[str, Any]] = []
-
-    dia = pass_summary.get("diarization", {})
-    for model_id, model_block in (dia.get("by_model") or {}).items():
-        if model_block.get("status") == "ok":
-            from_name = f"{pass_label}__diarization__{safe_model_id(model_id)}"
-            regions.extend(_diarization_to_ls(model_block.get("result"), from_name))
-
-    ast_block = pass_summary.get("ast", {})
-    if ast_block.get("status") == "ok":
-        regions.extend(
-            _classification_to_ls(
-                ast_block.get("result"),
-                f"{pass_label}__ast",
-                win_length=ast_win_length,
-                hop_length=ast_hop_length,
-            )
-        )
-
-    yam_block = pass_summary.get("yamnet", {})
-    if yam_block.get("status") == "ok":
-        regions.extend(
-            _classification_to_ls(
-                yam_block.get("result"),
-                f"{pass_label}__yamnet",
-                win_length=yamnet_win_length,
-                hop_length=yamnet_hop_length,
-            )
-        )
-
-    asr = pass_summary.get("asr", {})
-    alignment = pass_summary.get("alignment") or {}
-    align_by_model = alignment.get("by_model") or {}
-    for model_id, model_block in (asr.get("by_model") or {}).items():
-        if model_block.get("status") != "ok":
-            continue
-        from_name = f"{pass_label}__asr__{safe_model_id(model_id)}"
-        # Three-case branch:
-        # (a) ASR with native timestamps  -> use the ASR result for per-segment regions.
-        # (b) ASR text-only + successful alignment -> use the alignment result.
-        # (c) ASR text-only + alignment skipped or failed -> single full-audio TextArea.
-        align_block = align_by_model.get(model_id) or {}
-        asr_result = model_block.get("result")
-        if _asr_has_timestamps(asr_result):
-            regions.extend(_asr_to_ls(asr_result, from_name, duration_s))
-        elif align_block.get("status") == "ok":
-            # align_transcriptions returns List[List[ScriptLine | None]] —
-            # one inner list per input audio. We always pass a single audio,
-            # so unwrap to the inner segment list.
-            ar = align_block.get("result")
-            inner = ar[0] if isinstance(ar, list) and ar and isinstance(ar[0], list) else ar
-            regions.extend(_asr_to_ls(inner, from_name, duration_s))
-        else:
-            regions.extend(_asr_to_ls(asr_result, from_name, duration_s))
-
-    return {
-        "data": {
-            "audio": audio_uri,
-            "pass": pass_label,
-            "duration_s": duration_s,
-        },
-        "predictions": [
-            {
-                "model_version": f"senselab-analyze:{pass_label}",
-                "score": 1.0,
-                "result": regions,
-            }
-        ],
-    }
-
-
-def build_labelstudio_config(summary: dict[str, Any]) -> str:
-    """Build a Label Studio labeling-config XML matching this run's per-task tracks.
-
-    Generates one ``<Labels>`` control per (pass, analyzer, model) and one
-    ``<TextArea>`` control per (pass, asr_model). Speakers, scene labels,
-    and transcripts each become a stacked timeline annotation row.
-
-    The three-axis uncertainty tracks are appended downstream by
-    ``senselab.audio.workflows.audio_analysis.attach_uncertainty_tracks_to_ls``.
-    """
-    parts: list[str] = ["<View>", '  <Audio name="audio" value="$audio"/>']
-    seen_label_sets: dict[str, list[str]] = {}
-
-    for pass_label, pass_summary in summary.get("passes", {}).items():
-        # Diarization tracks: one per model, with that model's discovered speaker labels
-        dia_by_model = (pass_summary.get("diarization") or {}).get("by_model") or {}
-        for model_id, model_block in dia_by_model.items():
-            if model_block.get("status") != "ok":
-                continue
-            speakers = sorted({str(getattr(seg, "speaker", "?")) for seg in (model_block.get("result", [[]])[0] or [])})
-            if not speakers:
-                speakers = ["SPEAKER_00", "SPEAKER_01"]
-            seen_label_sets[f"{pass_label}__diarization__{safe_model_id(model_id)}"] = speakers
-
-        # AST scene labels
-        ast = pass_summary.get("ast") or {}
-        if ast.get("status") == "ok":
-            labels = _collect_classification_labels(ast.get("result"))
-            if labels:
-                seen_label_sets[f"{pass_label}__ast"] = sorted(labels)
-
-        # YAMNet scene labels
-        yam = pass_summary.get("yamnet") or {}
-        if yam.get("status") == "ok":
-            labels = _collect_classification_labels(yam.get("result"))
-            if labels:
-                seen_label_sets[f"{pass_label}__yamnet"] = sorted(labels)
-
-        # ASR: each model gets its own TextArea
-        asr_by_model = (pass_summary.get("asr") or {}).get("by_model") or {}
-        for model_id, model_block in asr_by_model.items():
-            if model_block.get("status") != "ok":
-                continue
-            from_name = f"{pass_label}__asr__{safe_model_id(model_id)}"
-            parts.append(
-                f'  <TextArea name="{from_name}" toName="audio" perRegion="true" '
-                f'editable="true" placeholder="ASR transcript ({model_id})"/>'
-            )
-
-    for from_name, label_values in sorted(seen_label_sets.items()):
-        parts.append(f'  <Labels name="{from_name}" toName="audio">')
-        for v in label_values:
-            v_escaped = v.replace('"', "&quot;")
-            parts.append(f'    <Label value="{v_escaped}"/>')
-        parts.append("  </Labels>")
-
-    parts.append("</View>")
-    return "\n".join(parts) + "\n"
-
-
-def _collect_classification_labels(result: Any) -> set[str]:  # noqa: ANN401
-    """Extract the union of label strings observed in a classify_audios output."""
-    labels: set[str] = set()
-    for window in _classification_windows(result):
-        if not isinstance(window, dict):
-            continue
-        for label in window.get("labels") or []:
-            if label:
-                labels.add(str(label))
-    return labels
 
 
 def _speech_presence_labels(args: argparse.Namespace) -> list[str]:
